requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
cinemagoer==2023.5.1
Django==5.2.1
djangorestframework==3.16.0
frozenlist==1.6.0
greenlet==3.2.2
idna==3.10
IMDbPY==2022.7.9
lxml==5.4.0
multidict==6.4.3
propcache==0.3.1
requests==2.32.3
soupsieve==2.7
SQLAlchemy==2.0.40
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
